[UPDATE][SCRIPT] Use hidden script folder in home
I'd rather not see my script folder most of the time, so let's make it a
hidden folder instead.

I also link it each time with stow instead of using a conditional on the
folder's existence, which is more robust when adding scripts after I
already finished an install and only want to update the symlinks.