[FIX][SHELL] Modify PATH at login only
It's cleaner to only export my modified PATH when I'm logging in,
instead of doing it at each shell launch.

Because Zsh and Bash don't use the same mechanism to let you know you
are in a login session, you gotta check both separately. But that
doesn't matter because lightdm doesn't launch `sh` as a login shell
anyway... So the condition is super messy.