[ADD][VIM] Shellcheck options for ALE
I want shellcheck to detect which shell is most appropriate on the `sh`
filetype, but explicitly use the `bash` dialect on `bash` and `zsh`
filetypes.

I also really don't like nullary conditions in shell, so I'll screen for
those too.

I might turn on the 'quote-safe-variables' check too, because I think
quoting variables is a good habit to take.