[ADD][GIT] Use vim-fugitive as default mergetool
I really like the way vim-fugitive show a merge conflict by having our
local index on the left (the target), and the remote changes on the
right (what we're trying to merge with our branch). The result is
obviously the file in the middle.

The d2o and d3o mappings take from local/remote for the current chunk.