[ADD][GIT] Alias m4 to vimdiff 4-window view
Sometimes it is useful to also have the base file accessible when
merging branches. This view still has the local file on the left and the
remote branch on the right, with the base in the middle and the result
at the bottom.

The d1o/d2o/d3o take from the local/base/remote for the current file.