[ADD][SHELL] Project-local environment w/ localrc
This little script will look at your current working directory and its
parent directories to source local configuration files (very useful for
Python  projects, or the way I build TC with a local PATH pre-pended).

It works relatively well but still is a very fragile solution that
depends on the order in which the shell has sourced its files. Because
the Termite script kept erasing the previous PROMPT_COMMAND value for
Bash I had to change to source it at the beginning of my configuration
file.