[FIX] Xterm window title for Zsh and Bash
The escape sequence to display a title for Xterm and Termite is the
same, and the export for the current directory does not seem to have any
negative impact on Xterm.

The VTE script uses `$HOSTNAME` which is a bashism for `$HOST`, so this
needs a correction in my shell profile.