[UPDATE][I3] Use keycode mappings for number row
If and when I want to switch to a different layout that uses numbers
without the shift key, unlike the traditional french layout, I still
want my shortcuts to work for switching workspaces.

I don't think any of my other shortcuts would have a problem switching
between AZERTY and QWERTY layout for example (well, selecting parent and
child containers would be switched, but I rarely need that
functionality and it is easy to switch on the fly).