[UPDATE][SCRIPT] Use expansion for PROMPT_COMMAND
It seems like the idiomatic way to update $PROMPT_COMMAND is to use
variable expansion to add a semi-colon and a space before the command
we're adding.

I also switched from prepending my commands to appending them, which
allows overriding behaviour, such as terminal titles which could be set
from system files.