[ADD] Tin configuration file
The Makefile has to be modified to handle tin in a particular fashion,
because Tin removes and then rewrite its whole configuration file after
closing, instead of only modifying the file.

A symbolic link to its whole configuration folder is therefore the only
manageable way to keep the configuration linked.